inline works
