+++ apps/roam/src/utils/useDiscourseData.ts
@@ -1,0 +1,200 @@
+import { DiscourseNode } from "./getDiscourseNodes";
+import { DiscourseRelation } from "./getDiscourseRelations";
+import { RelationDetails } from "./hyde";
+import getDiscourseContextResults from "./getDiscourseContextResults";
+import getPageUidByPageTitle from "roamjs-components/queries/getPageUidByPageTitle";
+import findDiscourseNode from "./findDiscourseNode";
+import getDiscourseRelations from "./getDiscourseRelations";
+import getDiscourseNodes from "./getDiscourseNodes";
+import { useCallback, useEffect, useMemo, useState } from "react";
+import getSettingValueFromTree from "roamjs-components/util/getSettingValueFromTree";
+import getBasicTreeByParentUid from "roamjs-components/queries/getBasicTreeByParentUid";
+import deriveDiscourseNodeAttribute from "./deriveDiscourseNodeAttribute";
+import normalizePageTitle from "roamjs-components/queries/normalizePageTitle";
+import { Result } from "roamjs-components/types/query-builder";
+
+export type DiscourseData = {
+  results: Awaited<ReturnType<typeof getDiscourseContextResults>>;
+  refs: number;
+};
+
+const cache: {
+  [tag: string]: DiscourseData;
+} = {};
+
+export const getOverlayInfo = async (
+  tag: string,
+  relations: ReturnType<typeof getDiscourseRelations>,
+): Promise<DiscourseData> => {
+  try {
+    if (cache[tag]) return cache[tag];
+
+    const nodes = getDiscourseNodes(relations);
+
+    const [results, refs] = await Promise.all([
+      getDiscourseContextResults({
+        uid: getPageUidByPageTitle(tag),
+        nodes,
+        relations,
+      }),
+      // @ts-ignore - backend to be added to roamjs-components
+      window.roamAlphaAPI.data.backend.q(
+        `[:find ?a :where [?b :node/title "${normalizePageTitle(
+          tag,
+        )}"] [?a :block/refs ?b]]`,
+      ),
+    ]);
+
+    return (cache[tag] = {
+      results,
+      refs: refs.length,
+    });
+  } catch (error) {
+    console.error(`Error getting overlay info for ${tag}:`, error);
+    return {
+      results: [],
+      refs: 0,
+    };
+  }
+};
+
+export const getAllReferencesOnPage = (pageTitle: string) => {
+  const referencedPages = window.roamAlphaAPI.data.q(
+    `[:find ?uid ?text
+      :where
+        [?page :node/title "${normalizePageTitle(pageTitle)}"]
+        [?b :block/page ?page]
+        [?b :block/refs ?refPage]
+        [?refPage :block/uid ?uid]
+        [?refPage :node/title ?text]]`,
+  );
+  return referencedPages.map(([uid, text]) => ({
+    uid,
+    text,
+  })) as Result[];
+};
+
+export const useDiscourseData = (tag: string) => {
+  console.log("useDiscourseData", tag);
+  const [loading, setLoading] = useState(true);
+  const [results, setResults] = useState<DiscourseData["results"]>([]);
+  const [refs, setRefs] = useState(0);
+  const [score, setScore] = useState<number | string>(0);
+
+  const tagUid = useMemo(() => getPageUidByPageTitle(tag), [tag]);
+  const discourseNode = useMemo(() => findDiscourseNode(tagUid), [tagUid]);
+  const relations = useMemo(() => getDiscourseRelations(), []);
+  const allNodes = useMemo(() => getDiscourseNodes(), []);
+
+  const getInfo = useCallback(
+    () =>
+      getOverlayInfo(tag, relations)
+        .then(({ refs, results }) => {
+          if (!discourseNode) return;
+          const attribute = getSettingValueFromTree({
+            tree: getBasicTreeByParentUid(discourseNode.type),
+            key: "Overlay",
+            defaultValue: "Overlay",
+          });
+          return deriveDiscourseNodeAttribute({
+            uid: tagUid,
+            attribute,
+          }).then((score) => {
+            setResults(results);
+            setRefs(refs);
+            setScore(score);
+          });
+        })
+        .finally(() => setLoading(false)),
+    [tag, tagUid, relations, discourseNode],
+  );
+
+  const refresh = useCallback(() => {
+    setLoading(true);
+    getInfo();
+  }, [getInfo]);
+
+  useEffect(() => {
+    getInfo();
+  }, [getInfo]);
+
+  const validRelations = useMemo(() => {
+    if (!discourseNode) return [];
+    const selfType = discourseNode.type;
+
+    return relations.filter(
+      (relation) =>
+        relation.source === selfType || relation.destination === selfType,
+    );
+  }, [relations, discourseNode]);
+
+  const uniqueRelationTypeTriplets = useMemo<RelationDetails[]>(() => {
+    if (!discourseNode) return [];
+    const relatedNodeType = discourseNode.type;
+
+    return validRelations.flatMap((relation) => {
+      const isSelfSource = relation.source === relatedNodeType;
+      const isSelfDestination = relation.destination === relatedNodeType;
+
+      let targetNodeType: string;
+      let currentRelationLabel: string;
+
+      if (isSelfSource) {
+        targetNodeType = relation.destination;
+        currentRelationLabel = relation.label;
+      } else if (isSelfDestination) {
+        targetNodeType = relation.source;
+        currentRelationLabel = relation.complement;
+      } else {
+        return [];
+      }
+
+      const identifiedTargetNode = allNodes.find(
+        (node) => node.type === targetNodeType,
+      );
+      if (!identifiedTargetNode) return [];
+
+      return [
+        {
+          relationLabel: currentRelationLabel,
+          relatedNodeText: identifiedTargetNode.text,
+          relatedNodeFormat: identifiedTargetNode.format,
+        },
+      ];
+    });
+  }, [validRelations, discourseNode, allNodes]);
+
+  const validTypes = useMemo(() => {
+    if (!discourseNode) return [];
+    const selfType = discourseNode.type;
+
+    const hasSelfRelation = validRelations.some(
+      (relation) =>
+        relation.source === selfType && relation.destination === selfType,
+    );
+    const types = Array.from(
+      new Set(
+        validRelations.flatMap((relation) => [
+          relation.source,
+          relation.destination,
+        ]),
+      ),
+    );
+    return hasSelfRelation ? types : types.filter((type) => type !== selfType);
+  }, [discourseNode, validRelations]);
+
+  return {
+    loading,
+    results,
+    refs,
+    score,
+    refresh,
+    tagUid,
+    discourseNode,
+    relations,
+    allNodes,
+    validRelations,
+    uniqueRelationTypeTriplets,
+    validTypes,
+  };
+};
